# api/openapi.yaml
openapi: 3.1.0
info:
  title: IOE Backend API
  version: 0.1.0
  description: >
    Single-tenant LMS backend. Authentication exchanges a Google ID token for an
    access token (Bearer, 15 minutes) and a rotating refresh token held in an
    HttpOnly cookie scoped to /v1/auth.
servers:
  - url: https://api.example.com
paths:
...
  /v1/courses:
    get:
      summary: List published courses, newest first
      description: >
        Public. An Authorization header is optional; when present it must carry a valid
        bearer token. Rate-limited per client IP.
      security: []
      parameters:
        - name: level
          in: query
          schema: { type: string, enum: [beginner, intermediate, advanced] }
        - name: price
          in: query
          schema: { type: string, enum: [free, paid] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: cursor
          in: query
          description: The next_cursor of the previous page. Opaque.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogPage" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      summary: Create a course
      security:
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCourseRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/users/{ownerID}/courses:
    get:
      summary: List an owner's courses
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/OwnerID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoursePage" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}:
    get:
      summary: Get a course. Courses without a live version are 404 to non-managers and anonymous callers.
      security:
        - {}
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/View"
      description: >
        Public for live courses; an Authorization header is optional and, when present,
        must be valid. Rate-limited per client IP. Managers get the working copy unless they
        ask for view=live; everyone else always gets the live version.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      summary: Update course details
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCourseDetailsRequest" }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/price:
    post:
      summary: Set the course price (NPR only)
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetPriceRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/publish:
    post:
      summary: >
        Publish the working copy as the next live version. Instructors publish approved courses
        only (409 approval_required otherwise); root admins publish without review.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/archive:
    post:
      summary: Archive a course
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/submit:
    post:
      summary: >
        Submit a draft or changes_requested course for review. Managers only. in_review and
        approved courses reject edits with 409 course_not_editable; the first edit to a
        published course reopens its working copy as a draft while readers keep the live version.
        Fails with 400 invalid_quiz_reference if a block references a missing quiz, or 400
        invalid_media_reference if a block references a missing media asset.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      responses:
        "204": { description: No Content }
        "400":
          description: Validation failure, invalid_quiz_reference, or invalid_media_reference
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/approve:
    post:
      summary: Approve a course in review. Root admins only.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApproveRequest" }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/request-changes:
    post:
      summary: Send a course in review back to its author. Root admins only.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewNoteRequest" }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/unpublish:
    post:
      summary: Take the live version off the catalog and send the course back to its author. Root admins only.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewNoteRequest" }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/reviews:
    get:
      summary: A course's review trail, oldest first. Managers only.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CourseReview" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/versions:
    get:
      summary: A course's published versions, newest first. Managers only.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CourseVersionSummary" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/versions/{versionNumber}:
    get:
      summary: One published version as a course. Managers only.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - name: versionNumber
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/discard-draft:
    post:
      summary: >
        Throw away the working copy's changes and restore it, lecture content included, from
        the live version. Managers only; the course must be live and draft or changes_requested
        (409 invalid_transition otherwise). Every restored lecture's content_revision moves on.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      responses:
        "200":
          description: The restored working copy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/in-review:
    get:
      summary: Courses awaiting review, longest waiting first. Root admins only.
      security:
        - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoursePage" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/settings/course-publishing:
    get:
      summary: The caller's publishing policy. independent for root admins, review_required for everyone else.
      security:
        - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoursePublishingSettings" }
        "401": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/sections:
    post:
      summary: Add a section
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/sections/{sectionID}:
    patch:
      summary: Rename a section
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/SectionID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      summary: Remove a section
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/SectionID"
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/lectures:
    post:
      summary: Add a lecture
      description: >
        A media_asset_id must name an uploaded asset of this course with the block's kind
        (video for video blocks, image for image blocks and flashcard cards); otherwise
        400 invalid_media_reference. A quiz_id must name a quiz of this lecture; otherwise
        400 invalid_quiz_reference.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddLectureRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/lectures/{lectureID}:
    patch:
      summary: Rename a lecture
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/LectureID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TitleRequest" }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      summary: Remove a lecture
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/LectureID"
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Secure-ioe_refresh
  parameters:
...
    CourseID:
      name: courseID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
    OwnerID:
      name: ownerID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
  responses:
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    MethodNotAllowed:
      description: Method Not Allowed
      headers:
        Allow:
          description: Allowed HTTP methods for the target resource
          required: true
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InternalError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    CoursePage:
      type: object
      required: [courses, total]
      properties:
        courses:
          type: array
          items: { $ref: "#/components/schemas/Course" }
        total: { type: integer }
    ApproveRequest:
      type: object
      additionalProperties: false
      properties:
        note: { type: string, maxLength: 4000 }
...
    CatalogPage:
      type: object
      required: [courses]
      properties:
        courses:
          type: array
          items: { $ref: "#/components/schemas/CourseSummary" }
        next_cursor:
          type: string
          description: Present only when another page exists. Opaque.
...
    CreateCourseRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        owner_id: { $ref: "#/components/schemas/ID" }
        title: { type: string }
        description: { type: string }
...